move Info types into pkg/logs/internal/status
